ILE: Port to the Windows 7.1 SDK build platform
Dual Ports configuration is now not providing any default
configuration anymore but this is an invalid implementation
anyway and will be replaced shortly by the real deal
(reading the Fusion desc file)

+++ DeviceAdapters/IntegratedLaserEngine/ActiveBlanking.cpp
@@ -110,11 +110,11 @@ int CActiveBlanking::OnValueChange( MM::PropertyBase * Prop, MM::ActionType Act 
         {
           if ( vRequestEnabled )
           {
-            MMILE_->LogMMMessage( "Enabling Active Blanking for line " + std::to_string( vLineIndex) + " FAILED" );
+            MMILE_->LogMMMessage( "Enabling Active Blanking for line " + std::to_string( static_cast< long long >( vLineIndex ) ) + " FAILED" );
           }
           else
           {
-            MMILE_->LogMMMessage( "Disabling Active Blanking for line " + std::to_string( vLineIndex ) + " FAILED" );
+            MMILE_->LogMMMessage( "Disabling Active Blanking for line " + std::to_string( static_cast< long long >( vLineIndex ) ) + " FAILED" );
           }
           return ERR_ACTIVEBLANKING_SET;
         }
